% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMB-class.R
\docType{class}
\name{HMB-class}
\alias{HMB-class}
\title{Class HMB}
\description{
Class \code{HMB} is the base class for the HMB-package
}
\seealso{
\code{\link{hmb}}, \code{\link{ghmb}}, \code{\link{tsmb}}, \code{\link{gtsmb}}
}
