% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethComp.R
\name{MethComp}
\alias{MethComp}
\title{Generate Report for Estimation Method Comparison}
\usage{
MethComp(
  n,
  SampFr,
  SampFrlim,
  EnLargeFactor,
  TileSide,
  PixelSize,
  meanY_S,
  sigmaY_S,
  meanX_S,
  sigmaX_S,
  meanZ_S,
  sigmaZ_S,
  meanX_Sa,
  sigmaX_Sa,
  meanZ_Sa,
  sigmaZ_Sa,
  meanZ_U,
  CorrYX,
  CorrYZ,
  CorrXZ,
  AutoCorrF = 0.8,
  AutoCorrG = 0.95,
  AutoCorrG_star = 0.8,
  AutoCorr = TRUE,
  module = "All",
  ylim = c(15, 15, 50, 20, 100),
  nameForestAtt = "AGB",
  nameRSdata = c("GEDI", "TanDEM-X"),
  nameStudyArea = " ",
  output_file = NULL
)
}
\arguments{
\item{n}{n}

\item{SampFr}{SampFr}

\item{SampFrlim}{SampFrlim}

\item{EnLargeFactor}{Enlarge factor}

\item{TileSide}{tile size}

\item{PixelSize}{pixel size}

\item{meanY_S}{mean Y of S}

\item{sigmaY_S}{sigma Y of S}

\item{meanX_S}{mean X of S}

\item{sigmaX_S}{sigma X of S}

\item{meanZ_S}{mean Z of Z}

\item{sigmaZ_S}{sigma Z of Z}

\item{meanX_Sa}{mean X of Sa}

\item{sigmaX_Sa}{sigma X of Sa}

\item{meanZ_Sa}{mean Z of Sa}

\item{sigmaZ_Sa}{sigma Z of Sa}

\item{meanZ_U}{mean Z of U}

\item{CorrYX}{correlation Y and X}

\item{CorrYZ}{correlation Y and Z}

\item{CorrXZ}{correlation X and Z}

\item{AutoCorrF}{autocorrelation F}

\item{AutoCorrG}{autocorrelation G}

\item{AutoCorrG_star}{autocorrelation G_star}

\item{AutoCorr}{autocorrelation}

\item{module}{module}

\item{ylim}{ylim}

\item{nameForestAtt}{name of forest attributes}

\item{nameRSdata}{name of remote sensing data}

\item{nameStudyArea}{name of study area}

\item{output_file}{output file location}
}
\description{
Generate Report for Estimation Method Comparison
}
