% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethComp.R
\name{MethComp}
\alias{MethComp}
\title{Generate Report for Estimation Method Comparison}
\usage{
MethComp(
  n,
  SampFr,
  SampFrlim,
  EnLargeFactor,
  TileSide,
  PixelSize,
  meanY_S,
  sigmaY_S,
  meanX_S,
  sigmaX_S,
  meanZ_S,
  sigmaZ_S,
  meanX_Sa,
  sigmaX_Sa,
  meanZ_Sa,
  sigmaZ_Sa,
  meanZ_U,
  CorrYX,
  CorrYZ,
  CorrXZ,
  AutoCorrF = 0.8,
  AutoCorrG = 0.95,
  AutoCorrG_star = 0.8,
  AutoCorr = TRUE,
  module = "All",
  ylim = c(15, 15, 50, 20, 100),
  nameForestAtt = "AGB",
  nameRSdata = c("GEDI", "TanDEM-X"),
  nameStudyArea = " ",
  output_file = NULL
)
}
\arguments{
\item{output_file}{}
}
\value{

}
\description{
Generate Report for Estimation Method Comparison
}
